+++ azext_iot/tests/iothub/jobs/test_iothub_jobs_int.py
@@ -5,6 +5,7 @@
 # --------------------------------------------------------------------------------------------
 
 import json
+
 from datetime import datetime, timedelta
 from azext_iot.tests import IoTLiveScenarioTest
 from azext_iot.tests.settings import DynamoSettings, ENV_SET_TEST_IOTHUB_BASIC